1) Change all entity's float columns & DTO's float fields to double; 2) add hourly forecast feature.

+++ app/src/main/java/io/farafonova/weatherapp/domain/model/CurrentForecastWithLocation.kt
@@ -5,12 +5,12 @@ data class CurrentForecastWithLocation(
     val temperature: Int,
     val feelsLikeTemperature: Int,
     val forecastTime: String,
-    val windSpeed: Float,
+    val windSpeed: Double,
     val windDegree: Int,
     val pressure: Int,
     val humidity: Int,
     val dewPoint: Int,
-    val uvi: Float,
+    val uvi: Double,
     val description: String,
     val location: Location
 )
