+++ app/src/main/java/io/farafonova/weatherapp/persistence/network/weather/WeatherService.kt
@@ -7,8 +7,8 @@ import retrofit2.http.Query
 interface WeatherService {
     @GET("onecall")
     suspend fun getWeather(
-        @Query("lat") latitude: Float,
-        @Query("lon") longitude: Float,
+        @Query("lat") latitude: Double,
+        @Query("lon") longitude: Double,
         @Query("lang") language: String,
         @Query("units") units: String,
         @Query("exclude") excludeParts: String?,
