+++ app/src/main/java/io/farafonova/weatherapp/ui/favorites/WeatherFavoritesRecyclerViewAdapter.kt
@@ -8,7 +8,7 @@ import androidx.recyclerview.widget.RecyclerView
 import io.farafonova.weatherapp.databinding.ListItemFavoriteLocationBinding
 import io.farafonova.weatherapp.domain.model.BriefCurrentForecastWithLocation
 
-class WeatherFavoritesRecyclerViewAdapter(private val onFavoriteClickListener: (Float, Float) -> Unit) :
+class WeatherFavoritesRecyclerViewAdapter(private val onFavoriteClickListener: (Double, Double) -> Unit) :
     ListAdapter<BriefCurrentForecastWithLocation, WeatherFavoritesViewHolder>(
         FavoritesWeatherEntryComparator()
     ) {
@@ -32,7 +32,7 @@ class WeatherFavoritesRecyclerViewAdapter(private val onFavoriteClickListener: (
 class WeatherFavoritesViewHolder(private val binding: ListItemFavoriteLocationBinding) :
     RecyclerView.ViewHolder(binding.root) {
 
-    fun bind(item: BriefCurrentForecastWithLocation, onClickListener: (Float, Float) -> Unit) {
+    fun bind(item: BriefCurrentForecastWithLocation, onClickListener: (Double, Double) -> Unit) {
         binding.weatherEntry = item
         binding.root.setOnClickListener {
             onClickListener(
